+++ src/com/crm/domain/easyui/DataGrid.java
@@ -10,14 +10,14 @@ import java.util.List;
  */
 public class DataGrid implements java.io.Serializable {
 
-	private Long total;	// 总记录数
+	private Integer total;	// 总记录数
 	private List rows;	// 每行记录
 	
-	public Long getTotal() {
+	public Integer getTotal() {
 		return total;
 	}
 
-	public void setTotal(Long total) {
+	public void setTotal(Integer total) {
 		this.total = total;
 	}
